@@ -17,66 +17,48 @@ use App\Mealz\MealBundle\Repository\DayRepositoryInterface;
 use App\Mealz\MealBundle\Repository\MealRepositoryInterface;
 use App\Mealz\MealBundle\Repository\ParticipantRepositoryInterface;
 use App\Mealz\MealBundle\Repository\SlotRepositoryInterface;
+use App\Mealz\MealBundle\Service\Doorman;
 use App\Mealz\MealBundle\Service\EventService;
 use App\Mealz\MealBundle\Service\Exception\ParticipationException;
 use App\Mealz\MealBundle\Service\ParticipationService;
 use App\Mealz\UserBundle\Entity\Profile;
 use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\ORM\EntityManagerInterface;
 use Exception;
 use Psr\Log\LoggerInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use stdClass;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
-/**
- * @Security("is_granted('ROLE_USER')")
- */
+#[IsGranted('ROLE_USER')]
 class ParticipantController extends BaseController
 {
-    private EventDispatcherInterface $eventDispatcher;
-    private EventService $eventSrv;
-    private ParticipationHelper $participationHelper;
-    private ParticipationService $participationSrv;
-    private MealRepositoryInterface $mealRepo;
-    private SlotRepositoryInterface $slotRepo;
-    private DayRepositoryInterface $dayRepo;
-    private ParticipantRepositoryInterface $participantRepo;
-    private LoggerInterface $logger;
-
     public function __construct(
-        EventDispatcherInterface $eventDispatcher,
-        EventService $eventSrv,
-        ParticipationHelper $participationHelper,
-        ParticipationService $participationSrv,
-        MealRepositoryInterface $mealRepo,
-        SlotRepositoryInterface $slotRepo,
-        DayRepositoryInterface $dayRepo,
-        ParticipantRepositoryInterface $participantRepo,
-        LoggerInterface $logger
+        private readonly EventDispatcherInterface $eventDispatcher,
+        private readonly EventService $eventSrv,
+        private readonly ParticipationHelper $participationHelper,
+        private readonly ParticipationService $participationSrv,
+        private readonly MealRepositoryInterface $mealRepo,
+        private readonly SlotRepositoryInterface $slotRepo,
+        private readonly DayRepositoryInterface $dayRepo,
+        private readonly ParticipantRepositoryInterface $participantRepo,
+        private readonly LoggerInterface $logger,
+        private readonly Doorman $doorman
     ) {
-        $this->eventDispatcher = $eventDispatcher;
-        $this->eventSrv = $eventSrv;
-        $this->participationSrv = $participationSrv;
-        $this->mealRepo = $mealRepo;
-        $this->slotRepo = $slotRepo;
-        $this->dayRepo = $dayRepo;
-        $this->participantRepo = $participantRepo;
-        $this->participationHelper = $participationHelper;
-        $this->logger = $logger;
     }
 
     /**
      * Lets the currently logged-in user either join a meal, or accept an already booked meal offered by a participant.
-     *
-     * @Security("is_granted('ROLE_USER')")
      */
+    #[IsGranted('ROLE_USER')]
     public function joinMeal(Request $request): JsonResponse
     {
         $profile = $this->getProfile();
         if (null === $profile) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $parameters = json_decode($request->getContent(), true);
@@ -91,11 +73,11 @@ class ParticipantController extends BaseController
         } catch (Exception $e) {
             $this->logException($e);
 
-            return new JsonResponse(['message' => '402: ' . $e->getMessage()], \Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR);
+            return new JsonResponse(['message' => '402: ' . $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
 
         if (null === $result) {
-            return new JsonResponse(['message' => '403: User is not allowed to join meal'], \Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR);
+            return new JsonResponse(['message' => '403: User is not allowed to join meal'], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
 
         $this->eventSrv->triggerJoinEvents($result['participant'], $result['offerer']);
@@ -116,34 +98,32 @@ class ParticipantController extends BaseController
                 'participantId' => $result['participant']->getId(),
                 'mealState' => $this->participationHelper->getMealState($meal),
             ],
-            \Symfony\Component\HttpFoundation\Response::HTTP_OK
+            Response::HTTP_OK
         );
     }
 
     /**
      * Lets the currently logged-in user either leave a meal, or put it up for offer.
-     *
-     * @Security("is_granted('ROLE_USER')")
      */
-    public function leaveMeal(Request $request): JsonResponse
+    #[IsGranted('ROLE_USER')]
+    public function leaveMeal(Request $request, EntityManagerInterface $entityManager): JsonResponse
     {
         $profile = $this->getProfile();
         if (null === $profile) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
         $parameters = json_decode($request->getContent(), true);
         $meal = $this->mealRepo->find($parameters['mealId']);
         $participant = $this->participationSrv->getParticipationByMealAndUser($meal, $profile);
 
-        if (false === $this->getDoorman()->isUserAllowedToLeave($meal)) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        if (false === $this->doorman->isUserAllowedToLeave($meal)) {
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         } elseif (null === $participant) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $participant->setCombinedDishes(null);
 
-        $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($participant);
         $entityManager->flush();
 
@@ -160,7 +140,7 @@ class ParticipantController extends BaseController
         return new JsonResponse([
             'slotId' => $slotID,
             'mealState' => $this->participationHelper->getMealState($meal),
-        ], \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+        ], Response::HTTP_OK);
     }
 
     public function updateCombinedMeal(
@@ -173,11 +153,11 @@ class ParticipantController extends BaseController
         try {
             $participationSrv->updateCombinedMeal($participant, $dishSlugs);
         } catch (ParticipationException $pex) {
-            return new JsonResponse(['message' => $pex->getMessage()], \Symfony\Component\HttpFoundation\Response::HTTP_UNPROCESSABLE_ENTITY);
+            return new JsonResponse(['message' => $pex->getMessage()], Response::HTTP_UNPROCESSABLE_ENTITY);
         } catch (Exception $exc) {
             $this->logException($exc);
 
-            return new JsonResponse(['message' => 'unexpected error'], \Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR);
+            return new JsonResponse(['message' => 'unexpected error'], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
 
         $this->eventDispatcher->dispatch(new ParticipationUpdateEvent($participant));
@@ -190,15 +170,14 @@ class ParticipantController extends BaseController
                     $participant->getCombinedDishes()->toArray()
                 ),
             ],
-            \Symfony\Component\HttpFoundation\Response::HTTP_OK
+            Response::HTTP_OK
         );
     }
 
     /**
      * Makes a booked meal by a participant to be available for taken over.
-     *
-     * @Security("is_granted('ROLE_USER')")
      */
+    #[IsGranted('ROLE_USER')]
     public function offerMeal(Request $request): JsonResponse
     {
         if (false === is_object($this->getUser())) {
@@ -207,13 +186,13 @@ class ParticipantController extends BaseController
 
         $participant = $this->getParticipantFromRequest($request);
         if (null === $participant) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
         $meal = $participant->getMeal();
 
         if ($this->getProfile() !== $participant->getProfile()
-            || false === $this->getDoorman()->isUserAllowedToSwap($meal)) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            || false === $this->doorman->isUserAllowedToSwap($meal)) {
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $participant->setOfferedAt(time());
@@ -225,25 +204,24 @@ class ParticipantController extends BaseController
         // trigger meal-offered event
         $this->eventDispatcher->dispatch(new MealOfferedEvent($participant));
 
-        return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+        return new JsonResponse(null, Response::HTTP_OK);
     }
 
     /**
      * Cancels an offered meal by a participant, so it can no longer be taken over by other users.
-     *
-     * @Security("is_granted('ROLE_USER')")
      */
+    #[IsGranted('ROLE_USER')]
     public function cancelOfferedMeal(Request $request): JsonResponse
     {
         $parameters = json_decode($request->getContent(), true);
 
         if (null === $parameters['mealId']) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $participant = $this->getParticipantFromRequest($request);
         if (null === $participant) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $participant->setOfferedAt(0);
@@ -267,9 +245,7 @@ class ParticipantController extends BaseController
         ]);
     }
 
-    /**
-     * @Security("is_granted('ROLE_KITCHEN_STAFF')")
-     */
+    #[IsGranted('ROLE_KITCHEN_STAFF')]
     public function getParticipationsForWeek(Week $week): JsonResponse
     {
         $days = $week->getDays();
@@ -288,12 +264,10 @@ class ParticipantController extends BaseController
             $response = $this->addParticipationInfo($response, $participants, $day);
         }
 
-        return new JsonResponse($response, \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+        return new JsonResponse($response, Response::HTTP_OK);
     }
 
-    /**
-     * @Security("is_granted('ROLE_KITCHEN_STAFF')")
-     */
+    #[IsGranted("ROLE_KITCHEN_STAFF")]
     public function add(Profile $profile, Meal $meal, Request $request): JsonResponse
     {
         $parameters = json_decode($request->getContent(), true);
@@ -326,17 +300,15 @@ class ParticipantController extends BaseController
                 'day' => $meal->getDay()->getId(),
                 'profile' => $profile->getUsername(),
                 'booked' => $participationData,
-            ], \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+            ], Response::HTTP_OK);
         } catch (Exception $e) {
             $this->logException($e);
 
-            return new JsonResponse(['message' => $e->getMessage()], \Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR);
+            return new JsonResponse(['message' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
-    /**
-     * @Security("is_granted('ROLE_KITCHEN_STAFF')")
-     */
+    #[IsGranted("ROLE_KITCHEN_STAFF")]
     public function remove(Profile $profile, Meal $meal): JsonResponse
     {
         try {
@@ -361,23 +333,21 @@ class ParticipantController extends BaseController
                 'day' => $meal->getDay()->getId(),
                 'profile' => $profile->getUsername(),
                 'booked' => $participationData,
-            ], \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+            ], Response::HTTP_OK);
         } catch (Exception $e) {
             $this->logException($e);
 
-            return new JsonResponse(['message' => $e->getMessage()], \Symfony\Component\HttpFoundation\Response::HTTP_INTERNAL_SERVER_ERROR);
+            return new JsonResponse(['message' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
-    /**
-     * @Security("is_granted('ROLE_KITCHEN_STAFF')")
-     */
+    #[IsGranted("ROLE_KITCHEN_STAFF")]
     public function getProfilesWithoutParticipation(Week $week): JsonResponse
     {
         $participations = $this->participantRepo->getParticipantsOnDays($week->getStartTime(), $week->getEndTime());
         $response = $this->participationHelper->getNonParticipatingProfilesByWeek($participations);
 
-        return new JsonResponse($response, \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+        return new JsonResponse($response, Response::HTTP_OK);
     }
 
     /**
@@ -387,15 +357,15 @@ class ParticipantController extends BaseController
     {
         $profile = $this->getProfile();
         if (null === $profile) {
-            return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+            return new JsonResponse(null, Response::HTTP_FORBIDDEN);
         }
 
         $participant = $meal->getParticipant($profile);
         if (null === $participant) {
-            return new JsonResponse(['message' => 'No participation found'], \Symfony\Component\HttpFoundation\Response::HTTP_NOT_FOUND);
+            return new JsonResponse(['message' => 'No participation found'], Response::HTTP_NOT_FOUND);
         }
 
-        return new JsonResponse($participant->getCombinedDishes()->toArray(), \Symfony\Component\HttpFoundation\Response::HTTP_OK);
+        return new JsonResponse($participant->getCombinedDishes()->toArray(), Response::HTTP_OK);
     }
 
     private function generateResponse(string $route, string $action, Participant $participant): JsonResponse
@@ -423,7 +393,7 @@ class ParticipantController extends BaseController
      */
     private function logAdd(Meal $meal, Participant $participant): void
     {
-        if (false === is_object($this->getDoorman()->isKitchenStaff())) {
+        if (false === is_object($this->doorman->isKitchenStaff())) {
             return;
         }
 
@@ -440,7 +410,7 @@ class ParticipantController extends BaseController
 
     private function logRemove(Meal $meal, Participant $participant): void
     {
-        if (true === $this->getDoorman()->isKitchenStaff()) {
+        if (true === $this->doorman->isKitchenStaff()) {
             $logger = $this->logger;
             $logger->info(
                 'admin removed {profile} from {meal} (Meal: {mealId})',
