@@ -8,46 +8,32 @@ use JsonSerializable;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
- * @ORM\Entity
- * @ORM\Table(name="Category")
  * @SuppressWarnings(PHPMD.CamelCasePropertyName)
  * @SuppressWarnings(PHPMD.CamelCaseVariableName)
  * @SuppressWarnings(PHPMD.CamelCaseParameterName)
  */
+#[ORM\Entity]
+#[ORM\Table(name: 'Category')]
 class Category implements JsonSerializable
 {
-    /**
-     * @var int
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\Column(type: 'integer')]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
+    private int $id;
 
-    /**
-     * @Gedmo\Slug(fields={"title_en"})
-     * @ORM\Column(length=128, unique=true)
-     */
+    #[Gedmo\Slug(fields: ['title_en'])]
+    #[ORM\Column(type: 'string', length: 128, unique: true)]
     private ?string $slug = null;
 
-    /**
-     * @Assert\NotBlank()
-     * @Assert\Length(max=255)
-     * @ORM\Column(type="string", length=255, nullable=FALSE)
-     *
-     * @var string
-     */
-    protected $title_en;
+    #[Assert\NotBlank]
+    #[Assert\Length(max: 255)]
+    #[ORM\Column(type: 'string', length: 255, nullable: false)]
+    protected string $title_en;
 
-    /**
-     * @Assert\NotBlank()
-     * @Assert\Length(max=255)
-     * @ORM\Column(type="string", length=255, nullable=FALSE)
-     *
-     * @var string
-     */
-    protected $title_de;
+    #[Assert\NotBlank]
+    #[Assert\Length(max: 255)]
+    #[ORM\Column(type: 'string', length: 255, nullable: false)]
+    protected string $title_de;
 
     protected string $currentLocale = 'en';
 
@@ -69,10 +55,7 @@ class Category implements JsonSerializable
         $this->slug = $slug;
     }
 
-    /**
-     * @return string
-     */
-    public function getTitleEn()
+    public function getTitleEn(): string
     {
         return $this->title_en;
     }
@@ -82,10 +65,7 @@ class Category implements JsonSerializable
         $this->title_en = $title;
     }
 
-    /**
-     * @return string
-     */
-    public function getTitleDe()
+    public function getTitleDe(): string
     {
         return $this->title_de;
     }
@@ -95,10 +75,7 @@ class Category implements JsonSerializable
         $this->title_de = $title;
     }
 
-    /**
-     * @return string
-     */
-    public function getTitle()
+    public function getTitle(): string
     {
         if ('de' === $this->currentLocale && $this->title_de) {
             return $this->getTitleDe();
@@ -115,14 +92,6 @@ class Category implements JsonSerializable
     public function setCurrentLocale(string $locale): void
     {
         $this->currentLocale = $locale;
-    }
-
-    /**
-     * @return string
-     */
-    public function getCurrentLocale()
-    {
-        return $this->currentLocale;
     }
 
     public function jsonSerialize(): array
