@@ -14,55 +14,41 @@ use JsonSerializable;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
- * @ORM\Entity
  * @MealBundleAssert\DishConstraint()
- * @ORM\Table(name="meal")
  */
+#[ORM\Entity]
+#[ORM\Table(name: 'meal')]
 class Meal implements JsonSerializable
 {
-    /**
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
+    #[ORM\Id]
+    #[ORM\Column(name: 'id', type: 'integer')]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
     private ?int $id = null;
 
-    /**
-     * @ORM\ManyToOne(targetEntity="Dish", cascade={"refresh"}, fetch="EAGER")
-     * @ORM\JoinColumn(name="dish_id", referencedColumnName="id")
-     */
+    #[ORM\ManyToOne(targetEntity: Dish::class, cascade: ['refresh'], fetch: 'EAGER')]
+    #[ORM\JoinColumn(name: 'dish_id', referencedColumnName: 'id')]
     private Dish $dish;
 
-    /**
-     * @Assert\NotBlank()
-     * @ORM\Column(type="decimal", precision=10, scale=4, nullable=FALSE)
-     */
+    #[Assert\NotBlank]
+    #[ORM\Column(type: 'decimal', precision: 10, scale: 4, nullable: false)]
     private float $price = 0.0;
 
-    /**
-     * @Assert\NotBlank()
-     * @ORM\Column(type="integer", nullable=FALSE, name="participation_limit")
-     */
+    #[Assert\NotBlank]
+    #[ORM\Column(name: 'participation_limit', type: 'integer', nullable: false)]
     private int $participationLimit = 0;
 
-    /**
-     * @ORM\ManyToOne(targetEntity="Day", inversedBy="meals")
-     * @ORM\JoinColumn(name="day", referencedColumnName="id")
-     */
+    #[ORM\ManyToOne(targetEntity: Day::class, inversedBy: 'meals')]
+    #[ORM\JoinColumn(name: 'day', referencedColumnName: 'id')]
     private Day $day;
 
-    /**
-     * @Assert\NotBlank()
-     * @Assert\Type(type="DateTime")
-     * @ORM\Column(type="datetime", nullable=FALSE)
-     */
+    #[Assert\NotBlank]
+    #[ORM\Column(type: 'datetime', nullable: false)]
     private DateTime $dateTime;
 
     /**
-     * @ORM\OneToMany(targetEntity="Participant", mappedBy="meal")
-     *
      * @psalm-var Collection<int, Participant>
      */
+    #[ORM\OneToMany(mappedBy: 'meal', targetEntity: Participant::class)]
     public ?Collection $participants = null;
 
     public function __construct(Dish $dish, Day $day)
@@ -188,23 +174,6 @@ class Meal implements JsonSerializable
         }
 
         return false;
-    }
-
-    /**
-     * @TODO don't load every participant object (raw sql query in repo?)
-     */
-    public function getTotalConfirmedParticipations(): int
-    {
-        $totalParticipation = 0;
-
-        foreach ($this->getParticipants() as $participation) {
-            /* @var Participant $participation */
-            if (true === $participation->isConfirmed()) {
-                ++$totalParticipation;
-            }
-        }
-
-        return $totalParticipation;
     }
 
     public function hasReachedParticipationLimit(): bool
