@@ -4,45 +4,28 @@ namespace App\Mealz\MealBundle\Entity;
 
 use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use JsonSerializable;
 
-/**
- * @ORM\Entity
- * @ORM\Table(name="week")
- */
+#[ORM\Entity]
+#[ORM\Table(name: 'week')]
 class Week extends AbstractMessage implements JsonSerializable
 {
-    /**
-     * @ORM\Column(type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     *
-     * @var int
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\Column(type: 'integer')]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
+    private int $id;
 
-    /**
-     * @ORM\Column(type="smallint", nullable=FALSE)
-     *
-     * @var int
-     */
-    private $year;
+    #[ORM\Column(type: 'smallint', nullable: false)]
+    private int $year;
 
-    /**
-     * @ORM\Column(type="smallint", nullable=FALSE)
-     *
-     * @var int
-     */
-    private $calendarWeek;
+    #[ORM\Column(type: 'smallint', nullable: false)]
+    private int $calendarWeek;
 
-    /**
-     * @ORM\OneToMany(targetEntity="Day", mappedBy="week", cascade={"all"})
-     * @ORM\OrderBy({"dateTime" = "ASC"})
-     *
-     * @var ArrayCollection
-     */
-    private $days;
+    #[ORM\OneToMany(mappedBy: 'week', targetEntity: Day::class, cascade: ['all'])]
+    #[ORM\OrderBy(['dateTime' => 'ASC'])]
+    private Collection $days;
 
     public function __construct()
     {
@@ -57,18 +40,12 @@ class Week extends AbstractMessage implements JsonSerializable
         return $this->id;
     }
 
-    /**
-     * @return ArrayCollection
-     */
-    public function getDays()
+    public function getDays(): Collection
     {
         return $this->days;
     }
 
-    /**
-     * @param ArrayCollection $days
-     */
-    public function setDays($days): void
+    public function setDays(Collection $days): void
     {
         $this->days = $days;
     }
@@ -81,26 +58,17 @@ class Week extends AbstractMessage implements JsonSerializable
         return $this->year;
     }
 
-    /**
-     * @param int $year
-     */
-    public function setYear($year): void
+    public function setYear(int $year): void
     {
         $this->year = $year;
     }
 
-    /**
-     * @return int
-     */
-    public function getCalendarWeek()
+    public function getCalendarWeek(): int
     {
         return $this->calendarWeek;
     }
 
-    /**
-     * @param int $calendarWeek
-     */
-    public function setCalendarWeek($calendarWeek): void
+    public function setCalendarWeek(int $calendarWeek): void
     {
         $this->calendarWeek = $calendarWeek;
     }
