@@ -6,9 +6,8 @@ use Doctrine\ORM\Mapping as ORM;
 
 /**
  * Model representing a Dish variation.
- *
- * @ORM\Entity
  */
+#[ORM\Entity]
 class DishVariation extends Dish
 {
 }
