@@ -7,9 +7,10 @@ use App\Mealz\MealBundle\Entity\Meal;
 use App\Mealz\MealBundle\Entity\Participant;
 use App\Mealz\UserBundle\Entity\Profile;
 use Symfony\Bundle\SecurityBundle\Security;
+use Symfony\Component\DependencyInjection\Attribute\Autoconfigure;
 
 /**
- * central business logic to determine if the currently logged in user is allowed to do a certain action.
+ * central business logic to determine if the currently logged-in user is allowed to do a certain action.
  *
  * For instance a user should not be allowed to register for a meal, that
  * starts in 10 minutes, but maybe a user with a special role ("cook") should be able
@@ -19,6 +20,7 @@ use Symfony\Bundle\SecurityBundle\Security;
  *
  * This logic should be accessible in controllers, templates and services.
  */
+#[Autoconfigure(lazy: true)]
 class Doorman
 {
     /**
@@ -26,7 +28,7 @@ class Doorman
      *
      * @see $this->hasAccessTo
      */
-    private const AT_MEAL_PARTICIPATION = 0;
+    private const int AT_MEAL_PARTICIPATION = 0;
 
     /**
      * Current timestamp.
